+++ project/3rdparty/include/live555/groupsock/groupsock_version.hh
@@ -19,7 +19,10 @@ along with this library; if not, write to the Free Software Foundation, Inc.,
 #ifndef _GROUPSOCK_VERSION_HH
 #define _GROUPSOCK_VERSION_HH
 
-#define GROUPSOCK_LIBRARY_VERSION_STRING	"2024.03.08"
-#define GROUPSOCK_LIBRARY_VERSION_INT		1709856000
+#define GROUPSOCK_LIBRARY_VERSION_STRING	"2024.04.14"
+#define GROUPSOCK_LIBRARY_VERSION_INT		1713052800
+
+extern char const* const groupsockLibraryVersionStr;
+extern int const groupsockLibraryVersionInt;
 
 #endif